Split configuration loading. Step 2
